Host fonts locally
This makes sure the Montserrat font is delivered along with the
html & styles of the generated documentation.

The practical upshots of this are that no more requests have to be
made to google which is a privacy improvement and the proper fonts
are now delivered even if the computer viewing the documentation is
offline.