Styles for headers inside user markdown
When markdown is used inside jsdoc, any headers are just emitted
as is; in practice this leads to bigger headers inside smaller
headers (h1 inside h2 in the readme), which does not look very good.

This commit introduces a special set of header styles for use inside
documentation, which makes the case above look great.